requirements.txt
google-generativeai
unstructured[all-docs]
langchain-core
python-dotenv
google-cloud-aiplatform
# openai

###query-search
llama-index-vector-stores-pinecone
llama-index-core
llama-index-embeddings-openai
llama-index-llms-openai

###
llama-index-llms-vertex

###UI
streamlit

##
tqdm

cohere
llama-index-postprocessor-cohere-rerank


##
firebase-admin
streamlit-feedback


jsonschema